requirements.txt
# Deep Learning Framework
tensorflow>=2.15.0
keras>=2.15.0

# Flask API
flask>=3.0.0
flask-cors>=4.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Augmentation
albumentations>=1.3.0

# Machine Learning Tools
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0

# Progress bars
tqdm>=4.65.0

# Model interpretability
grad-cam>=1.4.0

# Utilities
pyyaml>=6.0
